% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary.R
\name{annot_sRNAs}
\alias{annot_sRNAs}
\title{summary}
\usage{
annot_sRNAs(baerhunter_annot, original_annot, refseq_name)
}
\arguments{
\item{baerhunter_annot}{GFF3 genome annotation file produced by baerhunter}

\item{original_annot}{The available original annotation file}

\item{refseq_name}{Name of the reference sequence in the annotation file}
}
\value{
A list of 4 summary dataframes

A dataframe with detected sRNAs with the nearby genes as per naming conventions
}
\description{
summary function
}
