% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bam_processing.R
\name{major_features}
\alias{major_features}
\title{Extract major features from the annotation file}
\usage{
major_features(annotation_file, target_strand, original_sRNA_annotation)
}
\arguments{
\item{annotation_file}{GFF3 genome annotation file.}

\item{target_strand}{A character string indicating the strand. Supports two valies; '+' and '-'.}

\item{original_sRNA_annotation}{A string indicating how the biotype of pre-annotated ncRNA, which can be found in the attribute column.In case if the user does not know how the sRNA is annotated, it can be set as "unknown". In this case, all RNAs apart from tRNAs and rRNAs will be removed from the selection.}

\item{annot_file_directory}{The directory path for the annotation file (default is '.')}
}
\value{
A dataframe with the major features from a set strand.
}
\description{
The function extracts parent features only; it also excludes all non-coding RNAs that are already annotated in the file.
}
